// src/widgets/hero/types/hero.types.ts
import type { SiteButtonProps } from "@/shared/types/button.types";
import type { PropsWithClassName } from "@/shared/types/props-with-classname";
import type { PictureFormatType } from "@/shared/ui/picture";

export type HeroProps = {
	title?: string;
	description?: string;
	poster?: PictureFormatType;
	button?: SiteButtonProps;
} & PropsWithClassName;